Mercurial revision: 7b912ccc729b6a184c7e6cf7d460bb50301e7164

+++ replace/uuid_parse.c
@@ -226,7 +226,7 @@ uuid_parse(const char *in, uuid_t uu)
 		if (i== 36)
 			if (*cp == 0)
 				continue;
-		if (!isxdigit(*cp))
+		if (!isxdigit((int) *cp))
 			return -1;
 	}
 	uuid.time_low = strtoul(in, NULL, 16);
